fix build error for pull requests
When a pull request is made from an external repository, the variables
PYPI_TOKEN and TESTPYPI_TOKEN are unset in the build for security
reasons. Previously, having these variable unset would cause the build
to fail. Now, it is no longer an error for these variables to be unset.
Note that deployment will always be skipped if either of these variables
is unset.