install Python using pyenv on Mac OS
Installations using pyenv are cacheable and speed up the CI pipeline on
Mac OS. Python installations in $HOME/.pyenv/versions, which includes
all packages installed using pip, are now added to the cache. At the
install phase, we skip the installation of Python versions and their
associated pip packages that are already in the cache. We also remove
from the cache any old Python versions that will never be used again.