introduce build dependency autoupdater bot
Implemented using GitHub Actions, the dependency autoupdater bot
regularly checks for updates to buildcache, Cython, Python, and ICU. If
an update is found, the bot creates a new branch, adds a commit
containing an update to the dependency, and creates a pull request.

This bot removes the need to manually check for updates to the
build-time dependencies that are used in the Travis CI build. Also,
this bot, when combined with Travis CI's feature of building pull
requests, allow you to determine whether updating a dependency would
break something before you push such an update to the master branch.