specify which Python versions to use in manylinux builds
Previously, the build used all Python 3 versions present in the
container, which may or may not match the Python versions that we
support, especially as old versions of Python are removed and new ones
are added to the manylinux image. We now ensure that we build iknowpy
only for the versions of Python that we support.