autoupdate manylinux images
Tags for manylinux images are now fixed in the CD pipeline instead of
always getting the latest image. This allows for repeatable builds and
reduces the risk that a breaking change in manylinux images will cause a
build failure.

To keep manylinux images up to date, we use the autoupdater to create
pull requests for updating the images.