allow deployment to be triggered manually
If a build is triggered through the Travis UI with the following custom
configuration:

before_script: export FORCE_DEPLOY=1

then force a deploy of iknowpy, even if version.py has not been changed.

Allowing forced deployment is helpful in situations where deployment
fails for some reason after the version number has already been
incremented. Perhaps the change to version.py was made in a commit other
than the final commit in a push. This can happen if a merge commit is
created after the change to version.py. Or perhaps Travis CI or PyPI
encountered an internal error, resulting in a failed deployment or
partial deployment.