handle in-progress ICU release
When a new release of ICU occurs, installation packages tend to be
uploaded gradually rather than all at once. To prevent the autoupdater
from encountering an error trying to find non-existent files when the
release is not yet complete, we log a warning instead of failing the
entire job. The autoupdater will update ICU only when the .zip source
and the Win64 binaries are available.