separate build and deploy into their own stages
The previous configuration had the risk of partial deployment. If the
build succeeded on only some of the platforms during a deployment run,
deployment would occur for only those platforms. It makes more sense to
deploy only if builds succeed on all platform. With separate build and
deploy stages, this is now the case.