tag shared library name with unique hash on Windows
Building iknowpy on Windows now requires the pefile and
machomachomangler packages.

On Mac OS X and Linux, we are already renaming the iKnow and ICU
shared libraries when creating wheels. This avoids dependency hell by
ensuring that the wheel is self-contained and won't interfere with other
Python packages that use shared libraries with the same name. In fact,
the standard tooling for building wheels on Linux (particularly
auditwheel, a Linux-only tool) automatically renames dependent shared
libraries for this reason.

There is not (yet?) a similar tool for including DLLs in Windows wheels
that ensures that the wheel is self-contained and won't interfere with
other Python packages. So I implemented my own functionality that works
just like auditwheel. When iknowpy is built, each DLL dependency is
renamed by appending a unique hash based on the contents of the DLL. All
executables that require this DLL are then modified to point to the
new DLL name.