build wheels that work with multiple Python versions
Creating a wheel for every Python minor version on every platform takes
a lot of space on PyPI and limits the number of releases that we can do
before reaching the 10GiB PyPI limit and being forced to delete old
releases. It would be much more space-efficient to have a single wheel
per platform that works with all supported Python minor versions. This is
now possible by running `python setup.py merge` after creating a wheel
for each Python minor version.

This results in a 78% reduction in the size of a release.