run the PyPI size checker after deployment
Previously, there was a daily check of the size of the iknowpy project
on PyPI, which was unnecessarily frequent. Now, we check the size of
PyPI project after every deployment, and if the size is above a certain
threshold, an issue is created to alert us.

This threshold has also been increased from 8GiB to 9GiB because the
size of a release was greatly reduced by commit
7375fcaa3d2afcc121159be9681c620de23d1325.