increase manylinux ccache size from 350MiB to 500MiB
Even though we have not reached a cache size of 350MiB in our builds,
ccache is evicting entries in its cache and seems to be reducing the
performance of builds. ccache stores its files in 16 buckets and
enforces a size limit of 1/16th the total cache size limit to each
bucket. This means that if the buckets are unbalanced, entries can be
evicted from the cache even though we are not yet at the total size
limit. By increasing the cache size, I expect unnecessary evictions to
occur less frequently.