avoid XCode version conflicts
XCode version is now fixed to 12.2 to avoid any breaking changes that
has resulted from GitHub Actions's changing the Mac OS default XCode
version.

Cached ICU and Python installations on Mac OS now include the XCode
version as part of their keys. This means that if we change the version
of XCode used in the build, cached installations that were created with
a different XCode version will not be used.