provide force deploy option
Back when we were using Travis CI, we had a force deploy option to
handle situations where bumping the version number failed to trigger a
release. A similar feature is now available through the GitHub Actions
UI.

To trigger a release manually, go to the Actions tab of the repository,
click "CI" in the left panel, and click the "Run workflow" button to the
right.