support the --fast flag on Windows only
When `python setup.py install --fast` is invoked, we copy dependent
libraries as-is into the iknowpy installation location. This skips the
step where we edit the dependent libraries and improves performance in a
development scenario.

But it turns out that skipping the library editing step is safe to do on
Windows only due to how Windows searches for DLLs. On other platforms,
we need to edit the libraries during installation to ensure that they
are loadable by Python. So now, the --fast flag will be ignored on
non-Windows platforms.