check for updates to Python on Mac OS X on a schedule
Previously, the check for updates to Python on Mac OS X was triggered by
the CI workflow whenever an update to pyenv was detected. The problem
with this approach was that GitHub Actions's Mac OS machines are not all
updated at once, so a new version of pyenv would only sometimes be
available while GitHub Actions was rolling out an update. This caused
newer versions of Python to be available during some runs and not
others.

To avoid this issue, the dependency autoupdater now checks for updates
to Python on Mac OS X on the same schedule as for other dependencies. In
addition, both the autoupdater and the CI workflows run `brew update`
before installing pyenv to ensure consistency of the version of pyenv
being installed, thus ensuring that the same versions of Python are
available.