Remove superfluous preprocessing items
Recent changes in the normalizer have made
certain items in prepro.csv superfluous.
In Russian, removing the items from prepro.csv
had side-effects, which are solved by changes
in labels.csv and rules.csv. The positive effect of
these chanes is broader than just items that were
affected by prepro.csv.

Ref_testing reports in this commit reflect the
effect on the output. The reference will be
updated in a separate commit.