Check whether version number complies with PEP 440
a24b7f9da6b9914922f535327d0d19f158a981f0 exposed a problem where if the
version number does not comply with PEP 440, it may be normalized by
setuptools, but the build scripts are still looking for a wheel
containing the un-normalized version. This causes the build scripts to
fail to find the wheels that are created. To avoid this situation, we
now check the version number for PEP 440 compliance when building the
Python interface. This will allow non-compliant version number issues to
be caught locally.