Clean the cache regularly
GitHub Actions now has an API for cache management.

Once a day, we clear the GitHub Actions cache of single-use items that
are meant to be used only by 2 neighboring runs and will never be used
again. This keeps the cache clear of items that we know we will never
use again, reducing the chance that GitHub Actions will evict items that
we still need.